chore(test) - test infrastructure with docker
