+++ catchemall/web-api/src/index.js
@@ -1,37 +1,7 @@
-import { configDotenv } from 'dotenv';
-import express from 'express';
-import { validateConfig } from './config.js';
-import { createDbPool, seedDatabase } from './db/db.js';
-import { registerMiddleware } from './middleware/index.js';
-import { authRouter } from './routes/auth.router.js';
-import { pokemonRouter } from './routes/pokemon.router.js';
-import { userRouter } from './routes/user.router.js';
-import { createDailyPokemon, registerRefreshPokemonCron } from './business/dailyPokemon.js';
+import { bootstrap } from "./bootstrap.js";
 
-configDotenv();
+bootstrap().then(app =>
+  app.listen(4000, () => {
+    console.log(`catchemall api listening on port 4000`);
+  }));
 
-validateConfig();
-
-const app = express();
-const port = 4000;
-
-registerMiddleware(app);
-
-createDbPool();
-await seedDatabase();
-
-await createDailyPokemon();
-registerRefreshPokemonCron()
-app.get('/', (req, res) => {
-  res.send('Hello World!');
-});
-
-// Routes
-app.use("/pokemon", pokemonRouter);
-app.use("/auth", authRouter);
-app.use("/users", userRouter);
-
-
-app.listen(port, () => {
-  console.log(`Example app listening on port ${port}`);
-});
