+++ catchemall/web-api/test.sh
@@ -1,0 +1,1 @@
+docker compose -f docker-compose.test.yml up --build --exit-code-from api
