+++ jsbootcamp.code-workspace
@@ -10,6 +10,9 @@
     ]
   },
   "settings": {
+    "yaml.schemas": {
+      "https://json-schema.org/draft/2019-09/schema#":"docker-compose*.yml"
+    },
     "eslint.workingDirectories": [
       "./empty-react",
       "./learn-react",
